+++ CHANGELOG.md
@@ -16,7 +16,10 @@ All user visible changes to this project will be documented in this file. This p
 - UI:
     - Media panel:
         - Inability to drag-n-drop buttons in dock. ([#1654])
+- Windows:
+    - Unsupported devices displayed in output devices list for Windows 10. ([#1643])
 
+[#1643]: /../../pull/1643
 [#1654]: /../../pull/1654